[Gold I] Title: 제곱 ㄴㄴ 수, Time: 172 ms, Memory: 15976 KB -BaekjoonHub

+++ 백준/Gold/1016. 제곱 ㄴㄴ 수/제곱 ㄴㄴ 수.java
@@ -11,8 +11,8 @@ public class Main {
 
         boolean[] nums = new boolean[(int) (max - min + 1)];
 
-        for(long i = 2; i <= Math.sqrt(max); i++) {
-            long num = i * i;
+        for(int i = 2; i <= Math.sqrt(max); i++) {
+            long num = (long) i * i;
 
             for(long j = min / num; j <= max / num ; j++) {
                 if(num*j-min < 0) continue;
